+++ src/config.php
@@ -1,12 +1,12 @@
 <?php 
 
 /**
- * Email Entries config.php
+ * Email Content Editor config.php
  *
- * This file exists only as a template for the Email Entries settings.
+ * This file exists only as a template for the Email Content Editor settings.
  * It does nothing on its own.
  *
- * Don't edit this file, instead copy it to 'craft/config' as 'email-entries.php'
+ * Don't edit this file, instead copy it to 'craft/config' as 'email-content-editor.php'
  * and make your changes there to override default settings.
  *
  * Once copied to 'craft/config', this file will be multi-environment aware as